Development

+++ lib/screens/home/favorites/favorite_screen.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: prefer_final_fields
+// ignore_for_file: prefer_final_fields, use_build_context_synchronously
 
 import 'package:cache_manager/cache_manager.dart';
 import 'package:cached_network_image/cached_network_image.dart';
@@ -11,7 +11,6 @@ import 'package:healthy_buddy_mobile_app/screens/widgets/margin_height.dart';
 import 'package:healthy_buddy_mobile_app/shared/theme.dart';
 import 'package:provider/provider.dart';
 import 'package:sizer/sizer.dart';
-import 'package:url_launcher/url_launcher.dart';
 
 import '../../widgets/margin_width.dart';
 
@@ -25,12 +24,16 @@ class FavoriteScreen extends StatefulWidget {
 class _FavoriteScreenState extends State<FavoriteScreen> {
   int _currentIndex = 0;
   String? idUser;
-  List<bool> _selectedToogle = [true, false, false];
+  List<bool> _selectedToogle = [true, false];
   final List<String> _categoryLabel = [
     "Foodies",
-    "Sport",
     "MyDoc",
   ];
+
+  @override
+  void dispose() {
+    super.dispose();
+  }
 
   @override
   void initState() {
@@ -41,7 +44,6 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
         idUser = value;
         item.getFavFood(context: context, idUser: value);
         item.getFavMyDoc(context: context, idUser: value);
-        item.getFavSport(context: context, idUser: value);
       });
     });
   }
@@ -51,7 +53,6 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
     final item = Provider.of<FavoriteClass>(context);
     item.getFavFood(context: context, idUser: idUser ?? "");
     item.getFavMyDoc(context: context, idUser: idUser ?? "");
-    item.getFavSport(context: context, idUser: idUser ?? "");
     return Scaffold(
       backgroundColor: bgColor,
       body: SafeArea(
@@ -153,10 +154,8 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
       if (x == 0) {
         return item.food?.length ?? 0;
       } else if (x == 1) {
-        return item.sport?.length ?? 0;
-      } else if (x == 2) {
         return item.doc?.length ?? 0;
-      } else {
+      }  else {
         return 0;
       }
     }
@@ -194,9 +193,7 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
                   child: CachedNetworkImage(
                     imageUrl: _currentIndex == 0
                         ? '${item.food?[index].foodReceipt?.galleryPhoto?[0]}'
-                        : _currentIndex == 1
-                            ? '${item.sport?[index].sportExercise?.thumbnail}'
-                            : '${item.doc?[index].myDoc?.thumbnail}',
+                        : '${item.doc?[index].myDoc?.thumbnail}',
                     imageBuilder: (context, imageProvider) => Container(
                       decoration: BoxDecoration(
                         image: DecorationImage(
@@ -226,9 +223,7 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
                       child: Text(
                         _currentIndex == 0
                             ? '${item.food?[index].foodReceipt?.name}'
-                            : _currentIndex == 1
-                                ? '${item.sport?[index].sportExercise?.name}'
-                                : '${item.doc?[index].myDoc?.name}',
+                            : '${item.doc?[index].myDoc?.name}' ,
                         style: titleStyle.copyWith(fontSize: 12.sp),
                       ),
                     ),
@@ -237,9 +232,7 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
                       child: Text(
                         _currentIndex == 0
                             ? '${item.food?[index].foodReceipt?.description?.substring(0, 60)}...'
-                            : _currentIndex == 1
-                                ? '${item.sport?[index].sportExercise?.description?.substring(0, 60)}...'
-                                : '${item.doc?[index].myDoc?.description?.substring(0, 60)}...',
+                            : '${item.doc?[index].myDoc?.description?.substring(0, 60)}...',
                         style: regularStyle.copyWith(
                             color: greyTextColor, fontSize: 10.sp),
                       ),
@@ -247,24 +240,47 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
                     Row(
                       mainAxisAlignment: MainAxisAlignment.end,
                       children: [
-                        GestureDetector(
-                          onTap: () async {
-                            if (_currentIndex == 0) {
-                              await item.deleteFavFoodData(
-                                  id: item.food![index].id!, context: context);
-                            } else if (_currentIndex == 1) {
-                              await item.deleteFavSportData(
-                                  id: item.sport![index].id!, context: context);
-                            } else if (_currentIndex == 2) {
-                              await item.deleteFavMyDocData(
-                                  id: item.doc![index].id!, context: context);
-                            }
-                          },
-                          child: Icon(
-                            Icons.delete,
-                            color: blackColor,
-                          ),
-                        ),
+                        IconButton(
+                            onPressed: () {
+                              showDialog(
+                                context: context,
+                                builder: (context) => AlertDialog(
+                                  title: const Text('Konfirmasi'),
+                                  content: const Text(
+                                      'Apakah kamu yakin untuk menghapus item ini pada bagian favorite?'),
+                                  actions: <Widget>[
+                                    TextButton(
+                                      onPressed: () async {
+                                        try {
+                                          if (_currentIndex == 0) {
+                                            await item.deleteFavFoodData(
+                                                id: item.food![index].id!,
+                                                context: context);
+                                          } else if (_currentIndex == 1) {
+                                            await item.deleteFavMyDocData(
+                                                id: item.doc![index].id!,
+                                                context: context);
+                                          }
+                                        } catch (e) {
+                                          debugPrint(e.toString());
+                                        }
+                                        Navigator.pop(context);
+                                      },
+                                      child: const Text('Ya'),
+                                    ),
+                                    TextButton(
+                                      onPressed: () {
+                                        Navigator.pop(context);
+                                      },
+                                      child: const Text('Tidak'),
+                                    ),
+                                  ],
+                                ),
+                              );
+                            },
+                            icon: const Icon(
+                              Icons.delete,
+                            )),
                         MarginWidth(width: 4.h),
                         GestureDetector(
                           onTap: () async {
@@ -278,12 +294,6 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
                                 },
                               ));
                             } else if (_currentIndex == 1) {
-                              final sport = item.sport?[index].sportExercise;
-                              final url = Uri.parse(sport!.linkVideo!);
-                              if (await canLaunchUrl(url)) {
-                                await launchUrl(url);
-                              }
-                            } else if (_currentIndex == 2) {
                               final myDoc = item.doc?[index].myDoc;
                               Navigator.push(context, MaterialPageRoute(
                                 builder: (context) {
@@ -292,7 +302,7 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
                                   );
                                 },
                               ));
-                            }
+                            } 
                           },
                           child: Icon(
                             Icons.arrow_forward_outlined,
